src/pyproject.toml
[tool.poetry]
name = "duckling"
version = "0.1.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
sklearn = "^0.0"
numpy = "^1.19.2"
PyYAML = "^5.3.1"



[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
